splits up schemas, utils, tests
