+++ .circleci/config.yml
@@ -18,7 +18,7 @@ jobs:
           command: |
                   set -e
                   # set entrypoint like this so we can handle quotes
-                  docker run --entrypoint /bin/bash --read-only --tmpfs /tmp alleninstitutepika/ophys_etl_pipelines:${CIRCLE_SHA1} -c "/envs/suite2p/bin/python -m pytest -m 'suite2p_only or suite2p_also' --ignore-glob='*test_event_detection.py' /repos/ophys_etl/tests"
+                  docker run --entrypoint /bin/bash --read-only --tmpfs /tmp alleninstitutepika/ophys_etl_pipelines:${CIRCLE_SHA1} -c "/envs/suite2p/bin/python -m pytest -m 'suite2p_only or suite2p_also' --ignore='/repos/ophys_etl/tests/modules/event_detection' /repos/ophys_etl/tests"
                   docker run --entrypoint /bin/bash --read-only --tmpfs /tmp alleninstitutepika/ophys_etl_pipelines:${CIRCLE_SHA1} -c "/envs/event_detection/bin/python -m pytest -m 'event_detect_only' /repos/ophys_etl/tests"
                   docker run --entrypoint /bin/bash --read-only --tmpfs /tmp alleninstitutepika/ophys_etl_pipelines:${CIRCLE_SHA1} -c "/envs/ophys_etl/bin/python -m pytest -m 'not suite2p_only and not event_detect_only' /repos/ophys_etl/tests"
       - run:
