+++ src/ophys_etl/modules/event_detection/utils/utils.py
@@ -1,137 +1,14 @@
-import argschema
-import h5py
-import os
-import numpy as np
-import json
 import multiprocessing
-import warnings
-import marshmallow as mm
+import numpy as np
 from typing import Tuple
+from functools import partial
 from scipy.ndimage.filters import median_filter
 from scipy.stats import median_abs_deviation
-from functools import partial
 from FastLZeroSpikeInference import fast
-from ophys_etl.resources import event_decay_lookup_dict as decay_lookup
-from ophys_etl.schemas.fields import H5InputFile
 
 
 class EventDetectionException(Exception):
     pass
-
-
-class EventDetectionInputSchema(argschema.ArgSchema):
-    log_level = argschema.fields.Str(required=False, default="INFO")
-    movie_frame_rate_hz = argschema.fields.Float(
-        required=True,
-        description=("frame rate of the ophys video / trace. "
-                     "used to upsample slower rates to 31Hz."))
-    full_genotype = argschema.fields.Str(
-        required=False,
-        description=("full genotype of the specimen. Used to look "
-                     "up characteristic decay time from "
-                     "ophys_etl.resources.event_decay_time_lookup"))
-    decay_time = argschema.fields.Float(
-        required=False,
-        description="characteristic decay time [seconds]")
-    ophysdfftracefile = H5InputFile(
-        required=True,
-        description=("h5 file containing keys `data` (nROI x nframe) and "
-                     "`roi_names` (length = nROI) output by the df/F "
-                     "computation."))
-    valid_roi_ids = argschema.fields.List(
-        argschema.fields.Int,
-        required=True,
-        cli_as_single_argument=True,
-        description=("list of ROI ids that are valid, for which "
-                     "event detection will be performed."))
-    output_event_file = argschema.fields.OutputFile(
-        required=True,
-        description="location for output of event detection.")
-    n_parallel_workers = argschema.fields.Int(
-        required=False,
-        default=1,
-        description=("number of parallel workers. If set to -1 "
-                     "is set to multiprocessing.cpu_count()."))
-    noise_median_filter = argschema.fields.Float(
-        required=False,
-        default=1.0,
-        description=("median filter length used to detrend data "
-                     "during noise estimation [seconds]. Typically "
-                     "shorter than 'trace_median_filter'."))
-    trace_median_filter = argschema.fields.Float(
-        required=False,
-        default=3.2,
-        description=("median filter length used to detrend data "
-                     "before passing to FastLZero. [seconds]. Typically "
-                     "longer than 'noise_median_filter'."))
-    noise_multiplier = argschema.fields.Float(
-        required=False,
-        description=("manual specification of noise multiplier. If not "
-                     "provided, will be defaulted by `get_noise_multiplier` "
-                     "post_load below."))
-
-    @mm.post_load
-    def check_dff_h5_keys(self, data, **kwargs):
-        with h5py.File(data['ophysdfftracefile'], 'r') as f:
-            if 'roi_names' not in list(f.keys()):
-                raise EventDetectionException(
-                        f"DFF trace file {data['ophysdfftracefile']} "
-                        "does not have the key 'roi_names', which indicates "
-                        "it has come from an old version of creating "
-                        "DFF traces < April, 2019. Consider recreating the "
-                        "DFF file with a current version of that module.")
-        return data
-
-    @mm.post_load
-    def get_noise_multiplier(self, data, **kwargs):
-        if 'noise_multiplier' in data:
-            return data
-        # NOTE 11Hz value found empirically to behave well compared
-        # to 31Hz data sub-sampled to 11Hz.
-        val_11hz = 2.6
-        val_31hz = 2.0
-        if np.round(data['movie_frame_rate_hz'] / 11.0) == 1:
-            data['noise_multiplier'] = val_11hz
-        elif np.round(data['movie_frame_rate_hz'] / 31.0) == 1:
-            data['noise_multiplier'] = val_31hz
-        else:
-            raise EventDetectionException(
-                "You did not specify 'noise_multiplier'. In that case, the "
-                "multiplier is selected by 'movie_frame_rate_hz' but only "
-                "freqeuncies around 11Hz and 31Hz have specified values: "
-                f"({val_11hz}, {val_31hz}) respectively. You specified a "
-                f"'movie_frame_rate' of {data['movie_frame_rate_hz']}")
-        return data
-
-    @mm.post_load
-    def cpu_to_the_max(self, data, **kwargs):
-        if data['n_parallel_workers'] == -1:
-            data['n_parallel_workers'] = multiprocessing.cpu_count()
-        return data
-
-    @mm.post_load
-    def get_decay_time(self, data, **kwargs):
-        if ('full_genotype' not in data) & ('decay_time' not in data):
-            raise EventDetectionException(
-                    "Must provide either `decay_time` or `full_genotype` "
-                    "for decay time lookup. "
-                    "Available lookup values are "
-                    f"\n{json.dumps(decay_lookup, indent=2)}")
-        if 'full_genotype' in data:
-            if data['full_genotype'] not in decay_lookup:
-                raise EventDetectionException(
-                        f"{data['full_genotype']} not available. "
-                        "Available lookup values are "
-                        f"\n{json.dumps(decay_lookup, indent=2)}")
-            lookup = decay_lookup[data['full_genotype']]
-            if 'decay_time' in data:
-                warnings.warn("You specified a decay_time of "
-                              f"{data['decay_time']} but that is being "
-                              "overridden by a lookup by genotype to give "
-                              f"a decay time of {lookup}.")
-            data['decay_time'] = lookup
-        data['halflife'] = calculate_halflife(data['decay_time'])
-        return data
 
 
 def fast_lzero(penalty: float, dat: np.ndarray, gamma: float,
@@ -448,65 +325,3 @@ def get_events(traces: np.ndarray, noise_estimates: np.ndarray,
     events = np.array(events)
     lambdas = np.array(lambdas)
     return events, lambdas
-
-
-class EventDetection(argschema.ArgSchemaParser):
-    default_schema = EventDetectionInputSchema
-
-    def run(self):
-        self.logger.name = type(self).__name__
-        self.logger.setLevel(self.args.pop('log_level'))
-        ophys_etl_commit_sha = os.environ.get("OPHYS_ETL_COMMIT_SHA",
-                                              "local build")
-        self.logger.info(f"OPHYS_ETL_COMMIT_SHA: {ophys_etl_commit_sha}")
-
-        # read in the input file
-        with h5py.File(self.args['ophysdfftracefile'], 'r') as f:
-            roi_names = f['roi_names'][()].astype('int')
-            valid_roi_indices = np.argwhere(
-                    np.isin(
-                        roi_names,
-                        self.args['valid_roi_ids'])).flatten()
-            dff = f['data'][valid_roi_indices]
-        valid_roi_names = roi_names[valid_roi_indices]
-
-        empty_warning = None
-        if valid_roi_names.size == 0:
-            events = np.empty_like(dff)
-            lambdas = np.empty(0)
-            noise_stds = np.empty(0)
-            empty_warning = ("No valid ROIs in "
-                             f"{self.args['ophysdfftracefile']}. "
-                             "datasets in output file will be empty.")
-            self.logger.warn(empty_warning)
-        else:
-            # run FastLZeroSpikeInference
-            noise_filter_samples = int(self.args['noise_median_filter'] *
-                                       self.args['movie_frame_rate_hz'])
-            trace_filter_samples = int(self.args['trace_median_filter'] *
-                                       self.args['movie_frame_rate_hz'])
-            dff, noise_stds = estimate_noise_detrend(
-                    dff,
-                    noise_filter_size=noise_filter_samples,
-                    trace_filter_size=trace_filter_samples)
-            gamma = calculate_gamma(self.args['halflife'],
-                                    self.args['movie_frame_rate_hz'])
-            events, lambdas = get_events(
-                    traces=dff,
-                    noise_estimates=noise_stds * self.args['noise_multiplier'],
-                    gamma=gamma,
-                    ncpu=self.args['n_parallel_workers'])
-
-        with h5py.File(self.args['output_event_file'], "w") as f:
-            f.create_dataset("events", data=events)
-            f.create_dataset("roi_names", data=valid_roi_names)
-            f.create_dataset("noise_stds", data=noise_stds)
-            f.create_dataset("lambdas", data=lambdas)
-            if empty_warning:
-                f.create_dataset("warning", data=empty_warning)
-        self.logger.info(f"wrote {self.args['output_event_file']}")
-
-
-if __name__ == "__main__":  # pragma: no cover
-    ed = EventDetection()
-    ed.run()
